platform/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg2-binary==2.9.10
pgvector==0.3.6
pydantic==2.10.3
pydantic-settings==2.7.0
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.22
aiofiles==24.1.0
sentence-transformers==3.3.1
cryptography==46.0.5
websockets==14.1
httpx==0.28.1
Pillow==12.1.1
python-dotenv==1.0.1
apscheduler==3.10.4
